# [storage] Add serializeData helper function