@patrick-ogrady 2023-12-19
I'd be very very interested if you could benchmark the old/new functions for generating a key. I wholly agree the old way is confusing as heck.

I was thinking about using codec.Packer here instead of this approach (like we do for all other byte packing).