@martineckardt 2023-12-19
Ran some Benchmarks on the Performance:

### Original

```
Running tool: /usr/local/go/bin/go test -benchmem -run=^$ -bench ^BenchmarkSerializeOriginal$ github.com/ava-labs/hypersdk/examples/tokenvm/storage

goos: darwin
goarch: arm64
pkg: github.com/ava-labs/hypersdk/examples/tokenvm/storage
BenchmarkSerializeOriginal-12    	55650086	        21.27 ns/op	      64 B/op	       1 allocs/op
PASS
ok  	github.com/ava-labs/hypersdk/examples/tokenvm/storage	2.550s
```

### With helper

```
Running tool: /usr/local/go/bin/go test -benchmem -run=^$ -bench ^BenchmarkSerializeWithHelper$ github.com/ava-labs/hypersdk/examples/tokenvm/storage

goos: darwin
goarch: arm64
pkg: github.com/ava-labs/hypersdk/examples/tokenvm/storage
BenchmarkSerializeWithHelper-12    	 8424680	       136.9 ns/op	     192 B/op	       6 allocs/op
PASS
ok  	github.com/ava-labs/hypersdk/examples/tokenvm/storage	1.610s
```

###  Interpretation

- The original implementation performs significantly better in terms of speed, with a much lower time per iteration (21.27 ns/op) compared to the version with the helper function (135.2 ns/op).
- The original implementation also has lower memory allocations and fewer allocations per operation, indicating better memory efficiency.